Split PRCard into smaller components
